13b.1.1 Fix clippy warnings - format strings and unnecessary unwraps
Clean up 60+ clippy warnings across workspace to meet zero-warning quality gate for Linux CI validation.

Changes:
- Fix 51 uninlined_format_args warnings (use {var} instead of {}, var)
  * llmspell-kernel: 49 fixes across 20 files
  * llmspell-agents: 11 fixes across 14 files
  * llmspell-tools: 8 fixes across 4 files
  * llmspell-rag: 1 fix
  * llmspell-context: 2 fixes
  * llmspell-cli: 1 fix
- Fix 9 unnecessary_unwrap warnings in llmspell-templates
  * Replace is_some() + unwrap() pattern with if let
  * Applies to all 9 builtin template modules

Quality status: All minimal checks passing (format, clippy, compile, tracing)

+++ llmspell-tools/src/system/system_monitor.rs
@@ -460,8 +460,8 @@ impl SystemMonitorTool {
 
             if result == 0 {
                 let block_size = statvfs.f_bsize;
-                let total_blocks = u64::from(statvfs.f_blocks);
-                let available_blocks = u64::from(statvfs.f_bavail);
+                let total_blocks = statvfs.f_blocks;
+                let available_blocks = statvfs.f_bavail;
 
                 let total_bytes = total_blocks * block_size;
                 let available_bytes = available_blocks * block_size;
